+++ bika/lims/content/pricelist.py
@@ -4,6 +4,9 @@
 # Some rights reserved. See LICENSE.txt, AUTHORS.txt.
 
 from AccessControl import ClassSecurityInfo
+
+from decimal import Decimal
+
 from bika.lims import bikaMessageFactory as _
 from bika.lims.utils import t
 from bika.lims.browser.widgets.datetimewidget import DateTimeWidget
@@ -80,11 +83,11 @@ Field.widget.visible = True
 
 
 def apply_discount(price=None, discount=None):
-    return float(price) - (float(price) * float(discount)) / 100
+    return Decimal(price) - (Decimal(price) * Decimal(discount)) / 100
 
 
 def get_vat_amount(price, vat_perc):
-    return float(price) * float(vat_perc) / 100
+    return Decimal(price) * Decimal(vat_perc) / 100
 
 
 class PricelistLineItem(PersistentMapping):
@@ -144,8 +147,8 @@ def ObjectModifiedEventHandler(instance, event):
                     itemTitle = obj.Title()
                 cat = None
                 if obj.getPrice():
-                    price = float(obj.getPrice())
-                    totalprice = float(obj.getTotalPrice())
+                    price = Decimal(obj.getPrice())
+                    totalprice = Decimal(obj.getTotalPrice())
                     vat = totalprice - price
                 else:
                     price = 0
@@ -162,18 +165,18 @@ def ObjectModifiedEventHandler(instance, event):
                 #
                 cat = obj.getCategoryTitle()
                 if instance.getBulkDiscount():
-                        price = float(obj.getBulkPrice())
+                        price = Decimal(obj.getBulkPrice())
                         vat = get_vat_amount(price, obj.getVAT())
                         totalprice = price + vat
                 else:
                     if instance.getBulkPrice():
                         discount = instance.getBulkPrice()
-                        price = float(obj.getPrice())
+                        price = Decimal(obj.getPrice())
                         price = apply_discount(price, discount)
                         vat = get_vat_amount(price, obj.getVAT())
                         totalprice = price + vat
                     elif obj.getPrice():
-                        price = float(obj.getPrice())
+                        price = Decimal(obj.getPrice())
                         vat = get_vat_amount(price, obj.getVAT())
                         totalprice = price + vat
                     else:
