+++ bika/lims/tests/test_calculations.py
@@ -2,6 +2,7 @@
 #
 # Copyright 2011-2016 by it's authors.
 # Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+from decimal import Decimal
 
 from bika.lims import logger
 from bika.lims.content.analysis import Analysis
@@ -329,7 +330,7 @@ class TestCalculations(BikaFunctionalTestCase):
                         or an.isUpperDetectionLimit():
                         operator = an.getDetectionLimitOperand()
                         strres = f['analyses'][key].replace(operator, '')
-                        self.assertEqual(an.getResult(), str(float(strres)))
+                        self.assertEqual(an.getResult(), str(Decimal(strres)))
                     else:
                         self.assertEqual(an.getResult(), f['analyses'][key])
                 elif key == self.calcservice.getKeyword():
@@ -340,7 +341,7 @@ class TestCalculations(BikaFunctionalTestCase):
                 intermap = []
                 for i in interims:
                     if i['keyword'] in f['interims']:
-                        ival = float(f['interims'][i['keyword']])
+                        ival = Decimal(f['interims'][i['keyword']])
                         intermap.append({'keyword': i['keyword'],
                                         'value': ival,
                                         'title': i['title'],
@@ -354,7 +355,7 @@ class TestCalculations(BikaFunctionalTestCase):
 
             # Let's go.. calculate and check result
             calcanalysis.calculateResult(override=True, cascade=True)
-            self.assertEqual(float(calcanalysis.getResult()), float(f['exresult']))
+            self.assertEqual(Decimal(calcanalysis.getResult()), Decimal(f['exresult']))
 
     def test_calculation_fixed_precision(self):
         # Input results
@@ -402,7 +403,7 @@ class TestCalculations(BikaFunctionalTestCase):
                             or an.isUpperDetectionLimit():
                             operator = an.getDetectionLimitOperand()
                             strres = f['analyses'][key].replace(operator, '')
-                            self.assertEqual(an.getResult(), str(float(strres)))
+                            self.assertEqual(an.getResult(), str(Decimal(strres)))
                         else:
                             # The analysis' results have to be always strings
                             self.assertEqual(an.getResult(), str(f['analyses'][key]))
@@ -414,7 +415,7 @@ class TestCalculations(BikaFunctionalTestCase):
                     intermap = []
                     for i in interims:
                         if i['keyword'] in f['interims']:
-                            ival = float(f['interims'][i['keyword']])
+                            ival = Decimal(f['interims'][i['keyword']])
                             intermap.append({'keyword': i['keyword'],
                                             'value': ival,
                                             'title': i['title'],
@@ -478,7 +479,7 @@ class TestCalculations(BikaFunctionalTestCase):
                             or an.isUpperDetectionLimit():
                             operator = an.getDetectionLimitOperand()
                             strres = f['analyses'][key].replace(operator, '')
-                            self.assertEqual(an.getResult(), str(float(strres)))
+                            self.assertEqual(an.getResult(), str(Decimal(strres)))
                         else:
                             # The analysis' results have to be always strings
                             self.assertEqual(an.getResult(), str(f['analyses'][key]))
@@ -490,7 +491,7 @@ class TestCalculations(BikaFunctionalTestCase):
                     intermap = []
                     for i in interims:
                         if i['keyword'] in f['interims']:
-                            ival = float(f['interims'][i['keyword']])
+                            ival = Decimal(f['interims'][i['keyword']])
                             intermap.append({'keyword': i['keyword'],
                                             'value': ival,
                                             'title': i['title'],
